Add files via upload
Binning is the code used for trend removal and interpolation. NoamalFFT is used for spectral analysis and filtering, findpeaks is used to display peak values, and mark1 is used to mark the reciprocal of the horizontal axis at the graphical position clicked by the user's mouse, and display this value in red text above the clicked point. The newHilbert code is used for Hilbert transform.